# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API",
        "description": "Opentact API",
        "version": "1.0.0"
    },
    "host": "portal.opentact.org",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "paths": {
        "/api/conference/conference-list": {
            "get": {
                "tags": [
                    "conferences"
                ],
                "summary": "Get list of conferences",
                "responses": {
                    "200": {
                        "description": "List of conferences"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/conference/show/{conferenceName}": {
            "get": {
                "tags": [
                    "conferences"
                ],
                "summary": "Get list of conferences",
                "parameters": [
                    {
                        "name": "conferenceName",
                        "in": "path",
                        "description": "Conference name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conference"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/conference/list/{user_id}": {
            "get": {
                "tags": [
                    "files"
                ],
                "summary": "Return conference file list",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "APP user ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "conf_name",
                        "in": "query",
                        "description": "Conference name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File list"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/conference/file/{user_id}": {
            "get": {
                "tags": [
                    "files"
                ],
                "summary": "Return conference record file",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "APP user ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "conf_name",
                        "in": "query",
                        "description": "Conference name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "File name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/conference/add": {
            "post": {
                "tags": [
                    "conferences"
                ],
                "summary": "Create conference",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Conference name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "host_pin",
                        "in": "formData",
                        "description": "Host PIN",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "guest_pin",
                        "in": "formData",
                        "description": "Guest PIN",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "greeting_prompt",
                        "in": "formData",
                        "description": "Greeting prompt",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success result"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/conference/delete": {
            "post": {
                "tags": [
                    "conferences"
                ],
                "summary": "Delete conference",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Conference name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success result"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}